Runners simplified a bit.
